feature / логика фильтрации фильмов вынесена в хук useFilteredMovies
