+++ src/components/Movies/Movies.js
@@ -1,4 +1,4 @@
-import {useContext, useEffect, useState} from "react";
+import { useContext, useEffect, useState } from "react";
 
 import useCardsDisplay from "../../hooks/useCardsAmmount";
 
@@ -6,83 +6,41 @@ import SearchForm from "../SearchForm/SearchForm";
 import MoviesCardList from "../MoviesCardList/MoviesCardList";
 import Preloader from "../Preloader/Preloader";
 
-import {CurrentUserContext} from "../../contexts/CurrentUserContext";
+import { CurrentUserContext } from "../../contexts/CurrentUserContext";
 
-import {getAllDefaultMovies, shortFilter, wordFilter} from "../../utils/utils";
+import { getAllDefaultMovies } from "../../utils/utils";
+import useFilteredMovies from "../../hooks/useFilteredMovies";
 
-export default function Movies({getDefaultMovies}) {
+export default function Movies({ getDefaultMovies }) {
 
   const currentUser = useContext(CurrentUserContext);
   const storageKey = currentUser._id;
+
   const [allMovies, setAllMovies] = useState([]);
   const [loading, setLoading] = useState(false);
-  const [filterOn, setFilterOn] = useState(false);
-  const [keyWord, setKeyWord] = useState('');
-  const [foundMoviesArray, setFoundMoviesArray] = useState([]);
-  const [nothingFound, setNothingFound] = useState(false);
+
+  // Загрузка списка всех фильмов в начале работы страницы единожды
+  useEffect(() => {
+    const getAllMovies = async () => {
+      const movies = await getAllDefaultMovies(getDefaultMovies, setLoading);
+      setAllMovies(movies);
+    }
+    getAllMovies();
+  }, [getDefaultMovies]);
+
+  // Получение массива отфильтрованных карточек, значений и установщиков значений фильтров по ключевому слову и
+  // продолжительности, и отсутствия результатов поиска
+  const {
+    filterOn,
+    setFilterOn,
+    keyWord,
+    setKeyWord,
+    foundMoviesArray,
+    nothingFound
+  } = useFilteredMovies(allMovies, storageKey);
 
   // Получение текущего массива отображаемых карточек, обработчика нажатия на кнопку еще и состояния видимости кнопки ещё
   const { displayMovies, onAddBtnClick, addBtnVisible } = useCardsDisplay(foundMoviesArray);
-
-  // Восстановление конфигурации из хранилища
-  useEffect(() => {
-    if (localStorage.getItem(storageKey)) {
-
-      const config = JSON.parse(localStorage.getItem(storageKey));
-      setFilterOn(config.filterOn);
-      setKeyWord(config.keyWord);
-
-    }
-  }, [storageKey]);
-
-  // Сохранение конфигурацию при изменении фильтров
-  useEffect(() => {
-
-    localStorage.setItem(storageKey, JSON.stringify({
-      filterOn, keyWord,
-    }));
-
-  }, [keyWord, filterOn, storageKey]);
-
-  // Загрузка списка всех фильмов в начале работы страницы один раз
-  useEffect(() => {
-    const getAllMovies = async () => {
-
-      const movies = await getAllDefaultMovies(getDefaultMovies, setLoading);
-      setAllMovies(movies);
-
-    }
-
-    getAllMovies();
-
-  }, [getDefaultMovies]);
-
-  // Фильтрация списка фильмов по ключевому слову или ключевому слову и продолжительности
-  useEffect(() => {
-
-    if (keyWord !== '') {
-
-      let filteredMovies = keyWord ? allMovies.filter((movie) => wordFilter(keyWord, movie)) : allMovies;
-
-      if (filterOn) {
-        filteredMovies = filteredMovies.filter(movie => shortFilter(40, movie));
-      }
-
-      setFoundMoviesArray(filteredMovies);
-    }
-
-  }, [allMovies, keyWord, filterOn]);
-
-  // Установка наличия или отсутствия результатов поиска
-  useEffect(() => {
-
-    if ((keyWord !== '') && (foundMoviesArray.length === 0)) {
-
-      setNothingFound(true);
-
-    } else setNothingFound(false);
-
-  }, [foundMoviesArray]);
 
   return (<>
     <SearchForm filterOn={filterOn}
